pyproject.toml
[project]
name = "BlaBlaTeX"
version = "1.2.0"
description = "CLI tool to manage and use your personal LaTeX templates"
authors = [{ name = "Matti Bos", email = "matti.bos@gmx.de" }]
dependencies = ["typer[all]", "gitpython"]
requires-python = ">=3.7"

[project.scripts]
blablatex = "blablatex.cli:app"

[project.urls]
Repository = "https://github.com/KasperNr1/BlaBlaTeX"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"